refactor(test): split WAL recovery tests into separate files per test case
