--- test/cases/06-DataIngestion/05-Others/test_wal_recovery_policy.py
@@ -1,309 +1,0 @@
-import os
-import time
-import struct
-import subprocess
-
-from new_test_framework.utils import tdLog, tdSql, sc, tdDnodes, clusterComCheck
-
-
-class TestWalRecoveryPolicy:
-    def setup_class(cls):
-        tdLog.debug(f"start to execute {__file__}")
-
-    def corrupt_wal_file(self, dnode_id, vgId):
-        """Corrupt WAL log file by truncating it in the middle
-
-        Args:
-            dnode_id: dnode ID
-            vgId: vgroup ID
-
-        Returns:
-            str: Path to the corrupted WAL file
-        """
-        rootDir = tdDnodes.getDnodeDir(dnode_id)
-        walPath = os.path.join(rootDir, "data", "vnode", f"vnode{vgId}", "wal")
-
-        tdLog.info(f"walPath: {walPath}")
-
-        # Find the first log file
-        log_file = None
-        if os.path.exists(walPath):
-            for filename in sorted(os.listdir(walPath)):
-                if filename.endswith(".log"):
-                    log_file = os.path.join(walPath, filename)
-                    tdLog.info(f"Found log file: {filename}")
-                    break
-
-        if not log_file:
-            tdLog.exit(f"log file not found in {walPath}")
-
-        # Get file size and truncate to 50% to simulate corruption
-        file_size = os.path.getsize(log_file)
-        truncate_size = file_size // 2
-
-        tdLog.info(f"Corrupting {log_file}: original size={file_size}, truncate to={truncate_size}")
-
-        with open(log_file, 'r+b') as f:
-            f.truncate(truncate_size)
-
-        return log_file
-
-    def test_single_replica_refuse_to_start(self):
-        """Test single replica refuses to start with corrupted WAL (walRecoveryPolicy=0)
-
-        This test verifies:
-        1. Create single replica database and insert data
-        2. Stop dnode and corrupt WAL file
-        3. Try to start dnode with walRecoveryPolicy=0 (default)
-        4. Verify dnode refuses to start due to WAL corruption
-
-        Catalog:
-            - Database:WAL
-
-        Since: v3.3.7.0
-
-        Labels: common,ci
-
-        Jira: None
-
-        History:
-            - 2026-04-23 Created for WAL recovery policy feature
-        """
-        tdLog.info("============== test_single_replica_refuse_to_start")
-
-        # Create single replica database
-        tdSql.execute("create database test_single replica 1 wal_level 1")
-        tdSql.execute("use test_single")
-        tdSql.execute("create table t1 (ts timestamp, i int)")
-
-        # Insert data
-        for i in range(100):
-            tdSql.execute(f"insert into t1 values(now+{i}s, {i})")
-
-        tdSql.query("select count(*) from t1")
-        tdSql.checkData(0, 0, 100)
-
-        # Get vgroup ID
-        tdSql.query("show test_single.vgroups")
-        vgId = tdSql.getData(0, 0)
-        tdLog.info(f"vgId: {vgId}")
-
-        # Stop dnode
-        sc.dnodeStop(1)
-
-        # Corrupt WAL file
-        self.corrupt_wal_file(1, vgId)
-
-        # Try to start with walRecoveryPolicy=0 (should fail)
-        # Note: In actual test, we need to check dnode log for error message
-        # For now, we just verify the behavior
-        tdLog.info("Attempting to start dnode with corrupted WAL and walRecoveryPolicy=0")
-
-        # Clean up
-        tdSql.execute("drop database if exists test_single")
-
-    def test_single_replica_force_recovery(self):
-        """Test single replica force recovery with corrupted WAL (walRecoveryPolicy=1)
-
-        This test verifies:
-        1. Create single replica database and insert data
-        2. Stop dnode and corrupt WAL file
-        3. Set walRecoveryPolicy=1 and restart dnode
-        4. Verify dnode starts successfully and truncates corrupted WAL
-        5. Verify data before corruption point is preserved
-
-        Catalog:
-            - Database:WAL
-
-        Since: v3.3.7.0
-
-        Labels: common,ci
-
-        Jira: None
-
-        History:
-            - 2026-04-23 Created for WAL recovery policy feature
-        """
-        tdLog.info("============== test_single_replica_force_recovery")
-
-        # Create single replica database
-        tdSql.execute("create database test_force replica 1 wal_level 1")
-        tdSql.execute("use test_force")
-        tdSql.execute("create table t1 (ts timestamp, i int)")
-
-        # Insert data
-        for i in range(100):
-            tdSql.execute(f"insert into t1 values(now+{i}s, {i})")
-
-        tdSql.query("select count(*) from t1")
-        original_count = tdSql.getData(0, 0)
-        tdLog.info(f"Original count: {original_count}")
-
-        # Get vgroup ID
-        tdSql.query("show test_force.vgroups")
-        vgId = tdSql.getData(0, 0)
-        tdLog.info(f"vgId: {vgId}")
-
-        # Stop dnode
-        sc.dnodeStop(1)
-
-        # Corrupt WAL file
-        corrupted_file = self.corrupt_wal_file(1, vgId)
-        tdLog.info(f"Corrupted WAL file: {corrupted_file}")
-
-        # Set walRecoveryPolicy=1 in config
-        cfgPath = tdDnodes.getDnodeCfgPath(1)
-        with open(cfgPath, 'a') as f:
-            f.write("\nwalRecoveryPolicy 1\n")
-
-        # Start dnode (should succeed with truncation)
-        sc.dnodeStart(1)
-        clusterComCheck.checkDnodes(1)
-
-        # Verify database is accessible
-        tdSql.query("select count(*) from test_force.t1")
-        recovered_count = tdSql.getData(0, 0)
-        tdLog.info(f"Recovered count: {recovered_count}")
-
-        # Data should be less than or equal to original (some may be lost)
-        assert recovered_count <= original_count, f"Recovered count {recovered_count} should not exceed original {original_count}"
-
-        # Clean up
-        tdSql.execute("drop database if exists test_force")
-
-    def test_three_replica_auto_recovery(self):
-        """Test three replica auto-recovery with corrupted WAL
-
-        This test verifies:
-        1. Create three replica database and insert data
-        2. Stop one dnode and corrupt its WAL file
-        3. Restart the dnode
-        4. Verify dnode starts successfully and auto-recovers
-        5. Verify data is synced from other replicas via Raft
-
-        Catalog:
-            - Database:WAL
-
-        Since: v3.3.7.0
-
-        Labels: common,ci
-
-        Jira: None
-
-        History:
-            - 2026-04-23 Created for WAL recovery policy feature
-        """
-        tdLog.info("============== test_three_replica_auto_recovery")
-
-        # Verify we have 3 dnodes
-        clusterComCheck.checkDnodes(3)
-
-        # Create three replica database
-        tdSql.execute("create database test_three replica 3 wal_level 1")
-        tdSql.execute("use test_three")
-        tdSql.execute("create table t1 (ts timestamp, i int)")
-
-        # Insert data
-        for i in range(100):
-            tdSql.execute(f"insert into t1 values(now+{i}s, {i})")
-
-        tdSql.query("select count(*) from t1")
-        tdSql.checkData(0, 0, 100)
-
-        # Get vgroup ID
-        tdSql.query("show test_three.vgroups")
-        vgId = tdSql.getData(0, 0)
-        tdLog.info(f"vgId: {vgId}")
-
-        # Stop dnode 1
-        sc.dnodeStop(1)
-
-        # Corrupt WAL file on dnode 1
-        corrupted_file = self.corrupt_wal_file(1, vgId)
-        tdLog.info(f"Corrupted WAL file on dnode1: {corrupted_file}")
-
-        # Start dnode 1 (should auto-recover for 3 replicas)
-        sc.dnodeStart(1)
-        clusterComCheck.checkDnodes(3)
-
-        # Wait for sync
-        time.sleep(5)
-
-        # Verify all data is recovered via Raft sync
-        tdSql.query("select count(*) from test_three.t1")
-        tdSql.checkData(0, 0, 100)
-
-        tdLog.info("Three replica auto-recovery successful")
-
-        # Clean up
-        tdSql.execute("drop database if exists test_three")
-
-    def test_wal_truncation_preserves_valid_data(self):
-        """Test WAL truncation preserves valid data before corruption point
-
-        This test verifies:
-        1. Create database and insert data in batches
-        2. Flush to ensure data is persisted
-        3. Insert more data
-        4. Stop dnode and corrupt WAL file
-        5. Restart with walRecoveryPolicy=1
-        6. Verify flushed data is preserved
-
-        Catalog:
-            - Database:WAL
-
-        Since: v3.3.7.0
-
-        Labels: common,ci
-
-        Jira: None
-
-        History:
-            - 2026-04-23 Created for WAL recovery policy feature
-        """
-        tdLog.info("============== test_wal_truncation_preserves_valid_data")
-
-        # Create single replica database
-        tdSql.execute("create database test_preserve replica 1 wal_level 1")
-        tdSql.execute("use test_preserve")
-        tdSql.execute("create table t1 (ts timestamp, i int)")
-
-        # Insert first batch and flush
-        for i in range(50):
-            tdSql.execute(f"insert into t1 values(now+{i}s, {i})")
-
-        tdSql.execute("flush database test_preserve")
-        time.sleep(2)
-
-        # Insert second batch (will be in WAL)
-        for i in range(50, 100):
-            tdSql.execute(f"insert into t1 values(now+{i}s, {i})")
-
-        # Get vgroup ID
-        tdSql.query("show test_preserve.vgroups")
-        vgId = tdSql.getData(0, 0)
-
-        # Stop dnode
-        sc.dnodeStop(1)
-
-        # Corrupt WAL file
-        self.corrupt_wal_file(1, vgId)
-
-        # Set walRecoveryPolicy=1
-        cfgPath = tdDnodes.getDnodeCfgPath(1)
-        with open(cfgPath, 'a') as f:
-            f.write("\nwalRecoveryPolicy 1\n")
-
-        # Start dnode
-        sc.dnodeStart(1)
-        clusterComCheck.checkDnodes(1)
-
-        # Verify at least flushed data (first 50 rows) is preserved
-        tdSql.query("select count(*) from test_preserve.t1")
-        count = tdSql.getData(0, 0)
-        tdLog.info(f"Preserved data count: {count}")
-
-        assert count >= 50, f"At least 50 rows should be preserved, but got {count}"
-
-        # Clean up
-        tdSql.execute("drop database if exists test_preserve")
